% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_mdgrpub}
\alias{theme_mdgrpub}
\title{Personal Custom Theme with Titles}
\usage{
theme_mdgrpub(base_size = 12, base_family = "Lato")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Theme based on the \code{ggplot2 theme_bw} with modified font (UNIX use only). Thicker title case than plain \code{theme_mdgr.}
}
\examples{
library(ggplot2)
ggplot(mtcars) + geom_point(aes(wt, mpg)) + theme_mdgrpub()
}
